refactor: 페이지 및 컴포넌트 구조 재구성
