--- src/pages/AddItemPage.jsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function AddItemPage() {
-  return <div>AddItemPage</div>;
-}
