+++ src/pages/Anth/SignupPage.jsx
@@ -1,13 +1,16 @@
-import React, { useState } from "react";
+import { useState } from "react";
 import { Link } from "react-router-dom";
 import styled from "styled-components";
-import LogoHeader from "../components/LogoHeader";
-import Input from "../components/Input";
-import Button from "../components/Button";
-import SocialLogin from "../components/SocialLogin";
-import passwordHiddenIcon from "../assets/icon/password_hidden_icon.svg";
-import passwordVisibleIcon from "../assets/icon/password_visible_icon.svg";
-import validations from "../utils/validations";
+
+import validations from "@/utils/validations";
+
+import LogoHeader from "./components/LogoHeader";
+import Input from "@/components/Input";
+import Button from "@/components/Button";
+import SocialLogin from "./components/SocialLogin";
+
+import passwordHiddenIcon from "@/assets/icon/password_hidden_icon.svg";
+import passwordVisibleIcon from "@/assets/icon/password_visible_icon.svg";
 
 const Container = styled.div`
   width: 100%;
