+++ src/pages/Anth/components/LogoHeader.jsx
@@ -1,7 +1,6 @@
-import React from "react";
 import { Link } from "react-router-dom";
 import styled from "styled-components";
-import logo from "../assets/common/logo.svg";
+import logo from "@/assets/common/logo.svg";
 
 const Container = styled.div`
   margin: 60px 0 40px 0;
