+++ src/pages/MainPage/components/FeatureSection.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import styled from "styled-components";
 
 const Container = styled.div`
