+++ src/pages/MainPage/components/HeroSection.jsx
@@ -1,7 +1,7 @@
-import React from "react";
 import { Link } from "react-router-dom";
 import styled from "styled-components";
-import Button from "./Button";
+
+import Button from "@/components/Button";
 
 const Container = styled.div`
   width: 100%;
